handled confirmation before product delete

+++ src/app/dashboard/products/DeleteProduct.tsx
@@ -6,12 +6,14 @@ import { useRouter } from "next/navigation";
 const DeleteProduct = ({ item }: any) => {
   const router = useRouter();
   const handleClick = async () => {
-    alert(`sure to delete`)
-    const deletion = await axios.delete(`/api/products?id=${item}`)
+    const isConfirm = confirm(`sure to delete`)
+    if (isConfirm) {
+      const deletion = await axios.delete(`/api/products?id=${item}`)
 
-    if (deletion.status === 200) {
-      alert("Product Deleted");
-      router.push('/dashboard/products/')
+      if (deletion.status === 200) {
+        alert("Product Deleted");
+        router.push('/dashboard/products/')
+      }
     }
   }
